[Filter/Main] Support implicit dimension declaration
If a subplugin of tensor-filter has declared input/output dimension
with getInputDimenion/getOutputDimension, configure tensor-filter
with such values.

This prepares for configuring dimesions based on pad-cap negotiations as well.
(There are a few TODOs to fully support it)

This does about half of #144 .
Added a test case for implicit dimension declaration as well.

Signed-off-by: MyungJoo Ham <myungjoo.ham@samsung.com>