[Example Apps] Build example apps out of nnstreamer context
Build example_* apps in tizen build environment without nnstreamer
build environment as if nnstreamer is already in Tizen and
the apps are regular Tizen apps.

In order to do this:
- Add nnstreamer_testapp.spec: try gbs build --spec nnstreamer_testapp.spec to build test apps only
- Add INSTALL_EXAMPLE_APP option to install example apps optionally.
- Add tizen_app_build_example CMakeLists.txt to build example apps seperatedly.
- Added format dependencies on gstreamer-app-1.0

With this we can ensure that nnstreamer-based apps can be built completely without nnstreamer source codes and we can verify packaging of nnstreamer-devel.

Fixes #262 for Tizen.

Signed-off-by: MyungJoo Ham <myungjoo.ham@samsung.com>