[Tensorsave] Findtype file format definition.
To allow save tensors as files with its medata in tact,
we define "other/tensorsave" type for gstreamer.

It can be used to share tensors with non-gstreamer programs,
or used to create testcases for nnstreamer plugins.

To be completed, we need to implement load and save plugins
in gst/tensor_saveload directory.

Fixes #29

Signed-off-by: MyungJoo Ham <myungjoo.ham@samsung.com>