[Transform] Change parameter spec of 'mode' property to enum
This patch changes the existing parameter spec of string type
(g_param_spec) to the parameter spec of enum (g_param_enum), which gives
more flexibility by taking either strings or enum value, for the
arguments of the 'mode' property.  Moreover, this change makes the
gst-inspect-1.0 command show a list of supporting modes.

Signed-off-by: Wook Song <wook16.song@samsung.com>