[Android/ARM64] linker64: added a variable to maintain the custom linker64
This commit is to specify a variable to maintain the custom linker64.
This build script handles how to run the NNstreamer-based 64bit application
on Android 32bit platform. If the tart device is based on Android 64bit platform,
We do not need to use the custom linker64 repeatedly on the device.

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>