BUILD: Support nodebug option for gbs build
Fixed issue #1314.

This commit is to support optional flag to skip the generation of debug packages
to speed up a build time and to save a storage cost. In case that developers
have to do a debugging the source code with gdb, they may change the value of
the 'skipdebug' from 1 to 0.

**Changes**
V2:
 - Added "--define '_skip_debug_rpm 1'"
   : https://github.com/nnsuite/TAOS-CI/blob/master/ci/taos/plugins-base/pr-audit-build-tizen.sh#L84

V1:
 - Added optional flag to do not generate debug package

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>