[Tests/CustomFilters] Use default installation direcotory
Tests which require custom shared library does not consider the
tests using the unittest package (which means not the build time test).
They uses hardcoded path to indicate the location of the shared library.
In order to fix this problem, this patch makes the test use the default
installation directory when the test is running using the unittest
package.

Signed-off-by: Wook Song <wook16.song@samsung.com>