[Coverity/1087163] Fix dereference before null check
This intends to fix:
```
There may be a null pointer dereference, or else the comparison against null is unnecessary.

In ml_tensors_data_create: All paths that lead to this null pointer comparison already dereference the pointer earlier (CWE-476)
```

It appears that Coverity hates defensive coding.

This adds missing free on _data as well.

Signed-off-by: MyungJoo Ham <myungjoo.ham@samsung.com>