Policy Update: Test Tizen Features in ARM, always.
Because TAOS-CI skips unittests for ARM/ARM64 because of the latency,
we need to enable unittests for ARM-only tests unconditionally
if the test-target features are ARM-only.

Especially, this will help preventing regressions from other packages
in Tizen by enabling unit-tests in Tizen build (build.tizen.org) for
those selected features.

Changes in v2:
- Updated indentation (suggested by dongju.chae@samsung.com)

Signed-off-by: MyungJoo Ham <myungjoo.ham@samsung.com>