[TensorFilter] Add Add a callback of tensor filter for model update in runtime
This commit adds a callback, reloadModel(), of tensor filter for model
update in runtime. When the model property is updated, this callback
will be called if is-updatable is TRUE.

Also, it adds an event handler to update a model in the same pipeline.
On receiving a model update event, it calls set_property() of PROP_MODEL.

Signed-off-by: Dongju Chae <dongju.chae@samsung.com>