[TensorSink/UnitTest] Add a test option for sync property in TensorSink
This commit adds a test option, need_sync, for sync property in TensorSink.

Basically, GstBaseSink has a 'sync' property enabled by default, and
TensorSink also does. In this case, if the frame is later than max-lateness,
the sink will drop the buffer without calling the render method.

It can be a problem if some unittests (e.g., 'signals') expects the exact
number of received data. In slow system, if a signal callback is too slow,
some buffer can be dropped.

To fix this problem, this patch adds a test option for sync property and
set 'sync' property as FALSE by default to disable the buffer dropping
feature for 'signals' unittest.

Signed-off-by: Dongju Chae <dongju.chae@samsung.com>