[TFLite] add the datatype according to meson option
with the new version of TFLite, the new data types have been added. Since it's not supported and we cannot check the version of tensorflow at the build time, we have to set the datatype before building this package. I'd like to suggest adding a field at meson_option and allow user to set it.

Signed-off-by: HyoungJoo Ahn <hello.ahn@samsung.com>