[Repo/Sink] Consider a failure case of the render-buffer-function
A failure case of the function, gst_tensor_reposink_render_buffer, was
not carefully considered. Therefore, the pipeline becomes stuck when
gst_tensor_reposink_render_buffer fails. In order to avoid such situation,
this patch modifies the function to explicitly return GST_FLOW_ERROR when
a buffer could not be rendered.

Signed-off-by: Wook Song <wook16.song@samsung.com>