TAOS-CI: Updated the configuration files
New version of TAOS-CI is released on Mar-24-2020.
* Release note - https://github.com/nnsuite/TAOS-CI/releases/tag/v1.5.20200324

This commit is to update the existing configuration files of the nnstreamer
GitHub repository. The TAOS-CI can automatically load the configuraiton files
of the nnstreamer GitHub repository. Note that you do not must execute
the "update/merge" action on the configuration files that include the passwords
for security.

**Changelog**
  * Updated configuration files
  * Updated README.md file
  * Removed unnecessary files (Deprecated)

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>