[C-API] Allocate memory at once for tensor in tensors
This patch changes the semantics of memory allocation for tensor data.
Tensor data is now allocated for all the tensor in tensors.
This reduces the multiple calls to malloc/free to 1, thus reducing the latency.
The final location of the memory in the tensor data is set with the offset.
This updates the memory allocated ml_tensors_data_create() and memory allocated
internally in single API tensor filter.
Correspondongly ml_tensors_data_destroy is also correspondingly update.

Note that this does not update the memory allocation by the pipeline API elements.
The tensor filter in the pipeline API retains its original behavior.

This patch exposes 2 bugs :

1. Memory allocated by the tensor filters inside them (which support allocate_in_invoke)
is not freed appropriately. When using the single API, this memory is wrapped with ml_tensors_data
and freed wtih ml_tensors_data_destroy. However, these memories allocated by the tensor filter itself
must be destroyed with destroyer of the tensor filter extension (issue #2593).
This will be fixed in a different patch.

2. Single API sets the input/output tensors info provided by the user.
However, it sets all the info with empty string which is might not be even provided.
tensor_filter set_property() ignores empty properties for most of the times.
However it updates the number of tensors even with empty properties (issue #2594).
The failing unittest (nnstreamer_capi_singleshot.set_input_info_success_02) has been
fixed for now. A proper fix to the internals will be done in a different patch.

**Self evaluation:**
1. Build test: [x]Passed [ ]Failed [ ]Skipped
2. Run test: [x]Passed [ ]Failed [ ]Skipped

Signed-off-by: Parichay Kapoor <pk.kapoor@samsung.com>