[single] Resolve tensor filter memory leak
When tensor filter with allocate in invoke allocate the memory,
its memory should be destroyed with the filters destroy notify
However, this was ignored till now for single API (#2593)

This patch fixes this
Single handle now creates a list of each such data buffer
as well as the single handle is stored in the data buffer.
If data buffer is deallocated, then destroyNotify is called from the
single handle stored in it.
However, if single handle is closed before data buffer is freed,
then all the memory is freed from the list of the data buffers
stored in the single handle.

Resolves #2593

**Self evaluation:**
1. Build test: [x]Passed [ ]Failed [ ]Skipped
2. Run test: [x]Passed [ ]Failed [ ]Skipped

Signed-off-by: Parichay Kapoor <pk.kapoor@samsung.com>