[ext/tensor_filter] [patch 1/4] Initialize driver struct at compile-time
Initialize driver structs at compile-time by using designated
initializer.

This approach also improves readability and makes it easier to kepp
track of unimplemented APIs by partially forced to set fields manually
since g++ does not implement out-of-order (including skipped) designator.

Note that this commit fails to build, which shows the later point
mentioned. This is fixed in the later commit in this patch.

Signed-off-by: Bumsik Kim <k.bumsik@gmail.com>