[TensorRate] Implement the prototype codes for tensor_rate element
This patch implements the prototype codes for 'tensor_rate' element.

'tensor_rate' will include two methods to adjust a frame rate.
1) Either drop or duplicate frames similar to 'videorate' plugin
2) Propagate a specified frame-rate to upstream elements.

As a prototype, most of codes are reused from 'videorate' sources codes,
and still many things should be revised according to the tensor concept.

But, at least, its functionality was checked if we run this pipeline.

gst-launch-1.0 videotestsrc
   ! video/x-raw,width=640,height=480,framerate=15/1
   ! tensor_converter
   ! tensor_rate framerate=10/1
   ! tensor_decoder mode=direct_video
   ! videoconvert
   ! autovideosink

Signed-off-by: Dongju Chae <dongju.chae@samsung.com>