[tensor_filter_common] Fix a wrong example in accelerator prop spec
This patch fixes the wrong accelerator example in the prop spec.
This helps other developers to avoid confusions.
In actual tests, the true:(GPU,NPU,!CPU) format fails to be accepted so
that it switches to default mode. Also test cases do not cover such
format as well [1].

It also adds a note where to find available accelerator names.

[1]:
https://github.com/nnstreamer/nnstreamer/blob/main/tests/nnstreamer_filter_tensorflow_lite/runTest.sh#L108

Signed-off-by: Bumsik Kim <k.bumsik@gmail.com>