[Decoder/BoundingBox]: SSD: PP parameters can be set through option3
The post processing parameters used to decode a mobilenet SSD neural
network can be optionally set through option3, using ':' as parameter
delimiter.

The option3 definition scheme is, in order, the following:
- box priors location file (mandatory)
- Detection threshold (optional, default set to 0.5)
- Y box scale (optional, default set to 10.0)
- X box scale (optional, default set to 10.0)
- h box scale (optional, default set to 5.0)
- w box scale (optional, default set to 5.0)
- IOU box valid threshold (optional, default set to 0.5)

The default parameters value could be set through the following option3
example:

... option3=box-priors.txt:0.5:10.0:10.0:5.0:5.0:0.5 ...
or
... option3=box-priors.txt ...
or
... option3=box-priors.txt:::::: ...

It's possible to set only few values, using the default values for
those not specified through the command line.
You could specify respectively the detection and IOU thresholds to 0.65
and 0.6 with the option3 parameter as follow:

... option3=box-priors.txt:0.65:::::0.6 ...

Signed-off-by: Xavier Roumegue <xavier.roumegue@nxp.com>