[Query] Set TCP_NODELAY option to avoid buffering algorithm
Basically, the Nagle algorithm is enabled in TCP, which means that data
is buffered until there is a sufficient amount to send out. However, the
payload of Query should be sent as soon as possible, even if there is
only a small amount of data. This patch sets TCP_NODELAY option to avoid
buffering algorithm.

Signed-off-by: Sangjung Woo <sangjung.woo@samsung.com>