[filter/core] Add tensor filter latency reporting to the pipeline
Tensor filter executes inferences that may take long time depending on
the model complexity and the compute capabilities of the platform.
When inference duration is long, it happens that some downstream
buffers are dropped because associated timestamp is too far in the past.

It results in a shaky pipeline buffer rate and for instance occurences
of error messages from the sink when operated in live mode (sync=TRUE):
'There may be a timestamping problem, or this computer is too slow.'

Tensor filter implementation already has the option to computes latency
estimates when its element property latency=1 is set.

This change introduces a new property value latency_report=1 that will
when enabled use those latency duration estimates to:
- implement LATENCY query handling to include GstTensorFilter latency
estimate in the overall pipeline latency computation
- track filter latency estimates updates to notify the bus through a
LATENCY message that its latency has changed. It triggers a new round
of LATENCY query in the pipeline to take into account the updated filter
latency.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>