[tensor_converter] Remove a seemingly unnecessary caps attribute views=1
This fix makes tensor_converter compatible with vaapih264dec and adds a WARNING if incoming video caps views>1.

The need for this workaround and warning has been surfaced because of
a bug in vaapih264hdec https://gitlab.freedesktop.org/gstreamer/gstreamer/-/issues/1712
which provides `video/x-raw, ... multiview-mode=(string)mono, ... views=(int)2`.
The caps restriction of `views=(int)1` prohibits the successful negotiation.

This has been tested with a mono video but not with a truly
stereoscopic video which might need further handling in the buffer
processing.

Signed-off-by: Torsten Schulz <torsten.schulz@gmail.com>