[Test] Fix the test failure
After upgrading the version of GStreamer to 1.22.0, the test case of the
Tizen sensor filter failed. This is mainly because gst_parse_launch()
can return a non-NULL value even though the error is set. So error
should be checked first before using the return value of
gst_parse_launch(). This code fixes this bug.

Signed-off-by: Sangjung Woo <sangjung.woo@samsung.com>