[Filter] Allow dynamic framerate changes
One might want to dynamically change the framerate at any time while the
pipeline is running. This is prevented by code in `tensor_converter` and
`tensor_filter` that, upon reconfigure events, requre the new caps to be
'compatible' with the orignal caps on the sink pads. Requiring the caps
to have matching framerates seems unnecessary and too restrictive.

This patch removes that restriction from both `tensor_converter` and
`tensor_filter` elements.

A new unit test for this case has been added.

Signed-off-by: Torsten Schulz <torsten@eyepop.ai>