[query] Set tensor query caps when state changes from PAUSED to PLAYING
This patch makes to set tensor query caps, when state changes from PAUSED to PLAYING.
When the tensor_query_server is stopped, edge_handle is released.
So we need to set caps again when restarted.
If caps are not set, client connection is not possible after restart tensor_query_server.

Signed-off-by: Yelin Jeong <yelini.jeong@samsung.com>