[util] Remove unnecessary for loop execution
Remove unnecessary for loop execution in gst_tensors_info_free()
when tensor format is not '_NNS_TENSOR_FORMAT_STATIC'.

In gst_tensors_info_copy, if the tensor format is not '_NNS_TENSOR_FORMAT_STATIC',
only num_tensors and format information are updated without copying the tensor.
So only it needs to do gst_tensors_info_init. No need to run a for loop.

Signed-off-by: hyunil park <hyunil46.park@samsung.com>