bugfix : Memory management issue in XNNPACK delegate
g_assert is completely removed in release mode.
The assertion tensor_ptr->bytes == input[i].size (comparing tensor sizes)
is not a developer logic bug, but rather a data inconsistency that can occur at runtime.
This is an error that should be handled with an if statement, not g_assert.

Signed-off-by: hyunil park <hyunil46.park@samsung.com>